pyproject.toml
[project.urls]
homepage = "https://github.com/greenvla/GreenVLA"
issues = "https://github.com/greenvla/GreenVLA/issues"

[project]
name = "greenvla"
requires-python = ">=3.11"
keywords = ["robotics", "deep learning", "pytorch"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "cmake>=3.29.0.1",
    "datasets>=2.19.0",
    # "deepdiff>=7.0.1",
    "diffusers>=0.27.2",
    "draccus==0.11.5",
    "einops>=0.8.0",
    # "gdown>=5.1.0",
    # "h5py>=3.10.0",
    "huggingface-hub[hf-transfer,cli]>=0.27.1 ; python_version < '4.0'",
    "imageio[ffmpeg]>=2.34.0",
    "jsonlines>=4.0.0",
    # "numba>=0.59.0",
    "omegaconf>=2.3.0",
    "opencv-python-headless>=4.9.0",
    "packaging>=24.2",
    "av>=14.2.0",
    # "pymunk>=6.6.0,<7.0.0",
    # "pynput>=1.7.7",
    "pyzmq>=26.2.1",
    "rerun-sdk>=0.21.0",
    "termcolor>=2.4.0",
    "torch==2.7.1",
    "torchcodec==0.4.0; sys_platform != 'win32' and (sys_platform != 'linux' or (platform_machine != 'aarch64' and platform_machine != 'arm64' and platform_machine != 'armv7l')) and (sys_platform != 'darwin' or platform_machine != 'x86_64')",
    "torchvision==0.22.1",
    "safetensors==0.5.3",
    # "wandb>=0.16.3",
    "zarr>=2.17.0",
    "accelerate==1.4.0",
    "etils==1.12.2",
    "tyro==0.9.19",
    "importlib_resources==6.5.2",
    "sentencepiece==0.2.0",
    "transformers==4.57.3",
    "boto3==1.38.1",
    "types-boto3-s3==1.38.0",
    "tqdm_loggable==0.2",
    "numpydantic==1.6.8",
    "pydantic==2.10.6",
    "pydantic_core==2.27.2",
    "pytest==8.3.5",
    # "gcsfs==2024.12.0",
    "rich_argparse",
    "scipy==1.15.3",
    "hydra-core==1.3.2",
    "albumentations==2.0.8",
    "humanfriendly==10.0",
    "clearml==2.0.2",
    "matplotlib==3.10.6",
    "ffmpeg-python==0.2.0",
    "pip>=25.3",
    "uv>=0.9.26",
]

# Optional: uv sync --group notebook
[dependency-groups]
notebook = [
    "isort>=7.0.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
]

[tool.poetry]
requires-poetry = ">=2.1"

[tool.ruff]
line-length = 110
target-version = "py310"
exclude = ["tests/artifacts/**/*.safetensors"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "N", "B", "C4", "SIM"]

[tool.bandit]
exclude_dirs = [
    "tests",
    "benchmarks",
    "lerobot/common/datasets/push_dataset_to_hub",
    "lerobot/common/datasets/v2/convert_dataset_v1_to_v2",
    "lerobot/common/policies/pi0/conversion_scripts",
    "lerobot/scripts/push_dataset_to_hub.py",
]
skips = ["B101", "B311", "B404", "B603"]

[tool.typos]
default.extend-ignore-re = [
    "(?Rm)^.*(#|//)\\s*spellchecker:disable-line$",                      # spellchecker:disable-line
    "(?s)(#|//)\\s*spellchecker:off.*?\\n\\s*(#|//)\\s*spellchecker:on", # spellchecker:<on|off>
]
default.extend-ignore-identifiers-re = [
    # Add individual words here to ignore them
    "2nd",
    "pn",
    "ser",
    "ein",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
